requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
colorama>=0.4.6
prompt_toolkit>=3.0.43
rich>=13.7.0
psutil>=5.9.8
paramiko>=3.4.0
pandas>=2.1.4
python-dateutil>=2.8.2
ujson>=5.9.0
numpy>=1.24.0
scikit-learn>=1.3.0
setuptools>=69.0.3
pyinstaller>=6.2.0
